+++ packages/microsoft-security-adapter/src/client/connection.ts
@@ -10,7 +10,7 @@ import { AccountInfo } from '@salto-io/adapter-api'
 import { client as clientUtils } from '@salto-io/adapter-components'
 import { logger } from '@salto-io/logging'
 import { getAuthenticationBaseUrl } from './oauth'
-import { Credentials } from './types'
+import { Credentials } from '../auth'
 
 const log = logger(module)
