+++ packages/microsoft-security-adapter/src/definitions/fetch/entra/fetch.ts
@@ -837,18 +837,17 @@ const graphBetaCustomizations: FetchCustomizations = {
 
 // These types are always included in the fetch, regardless of the chosen services to manage
 const BASIC_ENTRA_TYPES = [GROUP_TYPE_NAME, GROUP_ADDITIONAL_DATA_TYPE_NAME]
-export const basicEntraCustomizations = Object.fromEntries(
-  Object.entries(graphV1Customizations).filter(([typeName, _customization]) => BASIC_ENTRA_TYPES.includes(typeName)),
-)
 
 export const createEntraCustomizations = ({
   entraExtended,
 }: {
   entraExtended: boolean
-}): Record<string, definitions.fetch.InstanceFetchApiDefinitions<Options>> =>
-  entraExtended
-    ? {
-        ...createCustomizationsWithBasePathForFetch(graphV1Customizations, GRAPH_V1_PATH),
-        ...createCustomizationsWithBasePathForFetch(graphBetaCustomizations, GRAPH_BETA_PATH),
-      }
-    : createCustomizationsWithBasePathForFetch(basicEntraCustomizations, GRAPH_V1_PATH)
+}): Record<string, definitions.fetch.InstanceFetchApiDefinitions<Options>> => {
+  const customizations = {
+    ...createCustomizationsWithBasePathForFetch(graphV1Customizations, GRAPH_V1_PATH),
+    ...createCustomizationsWithBasePathForFetch(graphBetaCustomizations, GRAPH_BETA_PATH),
+  }
+  return entraExtended
+    ? customizations
+    : _.pickBy(customizations, (_customization, typeName) => !BASIC_ENTRA_TYPES.includes(typeName))
+}
